# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Whitfield WMS API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/v1/inventory": {
   "get": {
    "tags": [
     "Inventory"
    ],
    "summary": "List Inventory",
    "description": "List inventory records.\n\nArgs:\n    warehouse_id (Optional[str]): Warehouse filter.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[InventoryResponse]: Inventory records.",
    "operationId": "list_inventory_v1_inventory_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "warehouse_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Warehouse Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InventoryResponse"
         },
         "title": "Response List Inventory V1 Inventory Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/inventory/stock/{upc}": {
   "get": {
    "tags": [
     "Inventory"
    ],
    "summary": "Stock By Upc",
    "description": "Return live stock for a product UPC at a warehouse.\n\nArgs:\n    upc (str): UPC barcode.\n    warehouse_id (str): Warehouse id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    StockLevelResponse: Live stock levels.",
    "operationId": "stock_by_upc_v1_inventory_stock__upc__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "upc",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Upc"
      }
     },
     {
      "name": "warehouse_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Warehouse Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StockLevelResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/inventory/low-stock": {
   "get": {
    "tags": [
     "Inventory"
    ],
    "summary": "Low Stock",
    "description": "List low-stock products.\n\nArgs:\n    warehouse_id (Optional[str]): Warehouse filter.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[StockLevelResponse]: Low-stock products.",
    "operationId": "low_stock_v1_inventory_low_stock_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "warehouse_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Warehouse Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/StockLevelResponse"
         },
         "title": "Response Low Stock V1 Inventory Low Stock Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/inventory/{inventory_id}": {
   "patch": {
    "tags": [
     "Inventory"
    ],
    "summary": "Adjust Inventory",
    "description": "Adjust inventory counts (admin/manager).\n\nArgs:\n    inventory_id (str): Inventory id.\n    request (InventoryAdjustRequest): Adjustment payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    InventoryResponse: Updated inventory.",
    "operationId": "adjust_inventory_v1_inventory__inventory_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "inventory_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Inventory Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InventoryAdjustRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InventoryResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v1/invoices": {
   "get": {
    "tags": [
     "Invoices"
    ],
    "summary": "List Invoices",
    "description": "List invoices (admin only).\n\nArgs:\n    seller_id (Optional[str]): Seller filter.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[InvoiceResponse]: Invoices.",
    "operationId": "list_invoices_v1_invoices_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "seller_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Seller Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InvoiceResponse"
         },
         "title": "Response List Invoices V1 Invoices Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/invoices/{invoice_id}": {
   "get": {
    "tags": [
     "Invoices"
    ],
    "summary": "Get Invoice",
    "description": "Fetch an invoice by id (admin only).\n\nArgs:\n    invoice_id (str): Invoice id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    InvoiceResponse: Invoice.",
    "operationId": "get_invoice_v1_invoices__invoice_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invoice_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Invoice Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvoiceResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/invoices/generate": {
   "post": {
    "tags": [
     "Invoices"
    ],
    "summary": "Generate Invoices",
    "description": "Generate invoices for a period (admin only).\n\nArgs:\n    year (int): Billing year.\n    month (int): Billing month.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[InvoiceResponse]: Generated invoices.",
    "operationId": "generate_invoices_v1_invoices_generate_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "year",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Year"
      }
     },
     {
      "name": "month",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "maximum": 12,
       "minimum": 1,
       "title": "Month"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InvoiceResponse"
         },
         "title": "Response Generate Invoices V1 Invoices Generate Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/notifications": {
   "get": {
    "tags": [
     "Notifications"
    ],
    "summary": "List Notifications",
    "description": "List notifications for the authenticated user.\n\nArgs:\n    unread_only (bool): Only unread notifications.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[NotificationResponse]: Notifications.",
    "operationId": "list_notifications_v1_notifications_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "unread_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Unread Only"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/NotificationResponse"
         },
         "title": "Response List Notifications V1 Notifications Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/notifications/{notification_id}/read": {
   "patch": {
    "tags": [
     "Notifications"
    ],
    "summary": "Mark Notification Read",
    "description": "Mark a notification as read.\n\nArgs:\n    notification_id (str): Notification id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    NotificationResponse: Updated notification.",
    "operationId": "mark_notification_read_v1_notifications__notification_id__read_patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "notification_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Notification Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InventoryAdjustRequest": {
    "properties": {
     "quantity_good": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quantity Good"
     },
     "quantity_damaged": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quantity Damaged"
     },
     "bin_location": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bin Location"
     }
    },
    "type": "object",
    "title": "InventoryAdjustRequest",
    "description": "Payload to adjust inventory stock counts.\n\n``quantity_good`` is the intended new good-stock count; reserved is preserved\nand available is recomputed as good minus reserved."
   },
...
   "InventoryResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "product_id": {
      "type": "string",
      "title": "Product Id"
     },
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "seller_id": {
      "type": "string",
      "title": "Seller Id"
     },
     "quantity_good": {
      "type": "integer",
      "title": "Quantity Good"
     },
     "quantity_damaged": {
      "type": "integer",
      "title": "Quantity Damaged"
     },
     "quantity_reserved": {
      "type": "integer",
      "title": "Quantity Reserved"
     },
     "quantity_available": {
      "type": "integer",
      "title": "Quantity Available"
     },
     "bin_location": {
      "type": "string",
      "title": "Bin Location"
     },
     "last_updated": {
      "type": "string",
      "title": "Last Updated"
     },
     "last_updated_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Updated By"
     }
    },
    "type": "object",
    "required": [
     "id",
     "product_id",
     "warehouse_id",
     "seller_id",
     "quantity_good",
     "quantity_damaged",
     "quantity_reserved",
     "quantity_available",
     "bin_location",
     "last_updated"
    ],
    "title": "InventoryResponse",
    "description": "Inventory representation."
   },
   "InvoiceLineItemResponse": {
    "properties": {
     "description": {
      "type": "string",
      "title": "Description"
     },
     "units": {
      "type": "number",
      "title": "Units"
     },
     "days": {
      "type": "number",
      "title": "Days"
     },
     "rate": {
      "type": "number",
      "title": "Rate"
     },
     "amount": {
      "type": "number",
      "title": "Amount"
     }
    },
    "type": "object",
    "required": [
     "description",
     "units",
     "days",
     "rate",
     "amount"
    ],
    "title": "InvoiceLineItemResponse",
    "description": "A charge line."
   },
   "InvoicePeriodResponse": {
    "properties": {
     "month": {
      "type": "integer",
      "title": "Month"
     },
     "year": {
      "type": "integer",
      "title": "Year"
     }
    },
    "type": "object",
    "required": [
     "month",
     "year"
    ],
    "title": "InvoicePeriodResponse",
    "description": "Billing period."
   },
   "InvoiceResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "invoice_ref": {
      "type": "string",
      "title": "Invoice Ref"
     },
     "seller_id": {
      "type": "string",
      "title": "Seller Id"
     },
     "seller_name": {
      "type": "string",
      "title": "Seller Name"
     },
     "period": {
      "$ref": "#/components/schemas/InvoicePeriodResponse"
     },
     "line_items": {
      "items": {
       "$ref": "#/components/schemas/InvoiceLineItemResponse"
      },
      "type": "array",
      "title": "Line Items"
     },
     "subtotal": {
      "type": "number",
      "title": "Subtotal"
     },
     "tax": {
      "type": "number",
      "title": "Tax"
     },
     "total": {
      "type": "number",
      "title": "Total"
     },
     "status": {
      "$ref": "#/components/schemas/InvoiceStatus"
     },
     "sent_at": {
      "type": "string",
      "title": "Sent At",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "id",
     "invoice_ref",
     "seller_id",
     "seller_name",
     "period",
     "line_items",
     "subtotal",
     "tax",
     "total",
     "status"
    ],
    "title": "InvoiceResponse",
    "description": "Invoice representation."
   },
   "InvoiceStatus": {
    "type": "string",
    "enum": [
     "draft",
     "sent"
    ],
    "title": "InvoiceStatus",
    "description": "Lifecycle status for an invoice."
   },
...
   "NotificationChannel": {
    "type": "string",
    "enum": [
     "email",
     "sms",
     "in_app",
     "push"
    ],
    "title": "NotificationChannel",
    "description": "Delivery channel for a notification."
   },
   "NotificationResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "recipient_type": {
      "$ref": "#/components/schemas/RecipientType"
     },
     "recipient_id": {
      "type": "string",
      "title": "Recipient Id"
     },
     "recipient_email": {
      "type": "string",
      "title": "Recipient Email"
     },
     "channel": {
      "$ref": "#/components/schemas/NotificationChannel"
     },
     "notification_type": {
      "$ref": "#/components/schemas/NotificationType"
     },
     "subject": {
      "type": "string",
      "title": "Subject"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "is_read": {
      "type": "boolean",
      "title": "Is Read"
     },
     "is_sent": {
      "type": "boolean",
      "title": "Is Sent"
     },
     "sent_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sent At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "recipient_type",
     "recipient_id",
     "recipient_email",
     "channel",
     "notification_type",
     "subject",
     "message",
     "is_read",
     "is_sent"
    ],
    "title": "NotificationResponse",
    "description": "Notification representation."
   },
   "NotificationType": {
    "type": "string",
    "enum": [
     "low_stock_alert",
     "shipment_received",
     "order_shipped",
     "return_processed",
     "invoice_ready",
     "daily_summary",
     "stale_order",
     "damage_report",
     "anomaly"
    ],
    "title": "NotificationType",
    "description": "Category of notification."
   },
...
   "RecipientType": {
    "type": "string",
    "enum": [
     "user",
     "seller"
    ],
    "title": "RecipientType",
    "description": "Type of notification recipient."
   },
...
   "StockLevelResponse": {
    "properties": {
     "product_id": {
      "type": "string",
      "title": "Product Id"
     },
     "product_name": {
      "type": "string",
      "title": "Product Name"
     },
     "upc_barcode": {
      "type": "string",
      "title": "Upc Barcode"
     },
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "quantity_good": {
      "type": "integer",
      "title": "Quantity Good"
     },
     "quantity_damaged": {
      "type": "integer",
      "title": "Quantity Damaged"
     },
     "quantity_reserved": {
      "type": "integer",
      "title": "Quantity Reserved"
     },
     "quantity_available": {
      "type": "integer",
      "title": "Quantity Available"
     },
     "bin_location": {
      "type": "string",
      "title": "Bin Location"
     }
    },
    "type": "object",
    "required": [
     "product_id",
     "product_name",
     "upc_barcode",
     "warehouse_id",
     "quantity_good",
     "quantity_damaged",
     "quantity_reserved",
     "quantity_available",
     "bin_location"
    ],
    "title": "StockLevelResponse",
    "description": "Live stock level for a product."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/v1/auth/login"
     }
    }
   }
  }
 }
}